backend_core/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-environ==0.11.2
pymongo==4.7.3
djongo==1.3.6
sqlparse==0.4.4
